requirements.txt
python-telegram-bot==20.7
openai==1.3.0
httpx>=0.25.0
sqlalchemy==2.0.23
feedparser==6.0.10
aiofiles==23.2.1
pandas==2.1.3
numpy==1.24.3
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
structlog==23.2.0
python-dateutil==2.8.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
orjson==3.9.10
diskcache==5.6.3
pydantic==2.5.0
pydantic-settings==2.1.0
click==8.1.7
rich==13.7.0
typer==0.9.0
fastapi==0.104.1
uvicorn==0.24.0